rename files to meet convention
